jobs/telemetry-dev-cycle/requirements.txt
click==8.0.4
pytest==6.0.2
pytest-black==0.3.11
pytest-flake8==1.0.6
requests==2.26.0
google-api-core==2.12.0
google-cloud-bigquery==2.30.1